fix:bug selling price in create item
