+++ src/app/(dashboard)/categories/(create)/create/page.tsx
@@ -1,6 +1,8 @@
+import { NextPage } from "next";
 import { CreateCategories } from "../_components/create-category";
+import { ReactElement } from "react";
 
-const Page = () => {
+const Page: NextPage = (): ReactElement => {
   return <CreateCategories />;
 };
 export default Page;
