+++ src/app/(dashboard)/categories/(edit)/_components/update-category.tsx
@@ -1,32 +1,45 @@
 "use client";
-
 import { Button, Form, Input, message } from "antd";
-import { useRouter } from "next/navigation";
-import { useCreateCategory } from "../_hooks/create-category.hook";
+import { useParams, useRouter } from "next/navigation";
+import { useEffect } from "react";
 import { TCategoryRequest } from "@/api/categories/type";
+import { useGetOneCategoryData } from "../../_hooks/get-one-category.hook";
+import { useUpdateCategory } from "../_hooks/update-category.hook";
 
-export const CreateCategories = () => {
+export const UpdateCategory = () => {
   const router = useRouter();
+  const params = useParams();
   const [form] = Form.useForm();
-  const { mutate } = useCreateCategory();
+  const { mutate } = useUpdateCategory();
+  const { data } = useGetOneCategoryData(params.id as string);
   const onFinish = async (values: TCategoryRequest) => {
     try {
-      mutate(values, {
-        onSuccess: () => {
-          message.success("Category created successfully!");
-          form.resetFields();
-          router.push("/categories");
-        },
-      });
+      mutate(
+        { id: params.id as string, payload: values },
+        {
+          onSuccess: () => {
+            message.success("Category update successfully!");
+            form.resetFields();
+            router.push("/categories");
+          },
+        }
+      );
     } catch (error) {
       throw new Error("Failed to create category");
       console.log(error);
     }
   };
+  useEffect(() => {
+    if (data) {
+      form.setFieldsValue({
+        name: data.name,
+      });
+    }
+  }, [data, form]);
   return (
     <div>
       <h2 style={{ textAlign: "center", marginBottom: "20px" }}>
-        Add New Category
+        Update Category
       </h2>
       <Form form={form} name="categories" onFinish={onFinish} layout="vertical">
         {" "}
