+++ src/app/(dashboard)/currencies/(edit)/[id]/update/page.tsx
