+++ src/app/(dashboard)/inventory/(create)/_hooks/create-inventory.hook.ts
