+++ src/app/(dashboard)/inventory/(create)/create/page.tsx
