+++ src/app/(dashboard)/inventory/(edit)/[id]/update/page.tsx
