+++ src/app/(dashboard)/inventory/(edit)/_hooks/update-inventory.hook.ts
